% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanr.R
\name{one_wasserstein_distance}
\alias{one_wasserstein_distance}
\title{One-dimensional Wasserstein distance}
\usage{
one_wasserstein_distance(left, right)
}
\arguments{
\item{left}{Numeric vector.}

\item{right}{Numeric vector.}
}
\value{
Numeric distance.
}
\description{
One-dimensional Wasserstein distance
}
\examples{
one_wasserstein_distance(c(0, 1, 2), c(1, 2, 3))
}
