% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanr.R
\name{scanr}
\alias{scanr}
\title{Scan your scripts folders for the functions you are looking
    for}
\usage{
scanr(pattern, path = "R", file_extension = "(r|R)$", ...)
}
\arguments{
\item{pattern}{a regular expression that matches what you are
looking for}

\item{path}{where to look for your R files}

\item{file_extension}{the extension for your R files (or other
regular expression) to match some files}

\item{...}{additional arguments to be passed to \code{grepl} to
match the regular expression for what you are looking for}
}
\value{
a tibble
}
\description{
Returns the functions for each file in
}
