% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanr.R
\name{ts_cusum}
\alias{ts_cusum}
\title{Localize a mean change with a CUSUM statistic}
\usage{
ts_cusum(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Integer split position.
}
\description{
Localize a mean change with a CUSUM statistic
}
\examples{
ts_cusum(c(rep(0, 5), rep(4, 5)))
}
