% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanr.R
\name{ts_wasserstein}
\alias{ts_wasserstein}
\title{Localize a distributional change with a Wasserstein statistic}
\usage{
ts_wasserstein(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
A list with \code{change_point} and \code{statistics}.
}
\description{
Localize a distributional change with a Wasserstein statistic
}
\examples{
result <- ts_wasserstein(c(rep(0, 5), rep(4, 5)))
result$change_point
}
