% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanr.R
\name{wasserstein_statistic}
\alias{wasserstein_statistic}
\title{One-dimensional Wasserstein distance}
\usage{
wasserstein_statistic(left, right)
}
\arguments{
\item{left}{Numeric vector.}

\item{right}{Numeric vector.}
}
\value{
Numeric distance.
}
\description{
One-dimensional Wasserstein distance
}
